Testung by addung a clock dyno
