+++ app.py
@@ -1,17 +1,10 @@
 from flask import Flask, request
 from twilio.twiml.messaging_response import MessagingResponse
-from apscheduler.schedulers.background import BackgroundScheduler
-from apscheduler.triggers.cron import CronTrigger
-from reddit import Message as RMessage
-import time
-import datetime
+
 
 app = Flask(__name__)
 
-def startscheduling():
-    scheduler = BackgroundScheduler()
-    scheduler.add_job(RMessage, trigger = CronTrigger(minute='*/01'))
-    scheduler.start()
+
 
 @app.route('/inb', methods=['GET', 'POST'])
 def reply():
@@ -23,5 +16,5 @@ def reply():
     return str(resp)
 
 if __name__ == '__main__':
-    startscheduling()
+#    startscheduling()
     app.run()
